requirements.txt
django 3.7.9,
django-colorfield 0.4.2,
Pillow 8.3.2,
djangorestframework 3.12.4,
djangorestframework-simplejwt 4.8.0,
